back/requirements.txt
aws-cdk.core
aws-cdk.aws-appsync
aws-cdk.aws-dynamodb
aws-cdk.aws-iam
aws-cdk.aws.cognito
configparser